Refactor email validator, remove remindAt creation in attendee
