+++ app/src/main/java/com/example/tasky/authentication/di/AuthenticationModule.kt
@@ -1,13 +1,13 @@
 package com.example.tasky.authentication.di
 
-import com.example.tasky.authentication.data.EmailValidatorImpl
 import com.example.tasky.authentication.data.local.UserCacheImpl
 import com.example.tasky.authentication.data.remote.AuthenticationRepositoryImpl
 import com.example.tasky.authentication.data.remote.login.LoginService
 import com.example.tasky.authentication.data.remote.signup.SignUpService
 import com.example.tasky.authentication.domain.AuthenticationRepository
-import com.example.tasky.authentication.domain.EmailValidator
 import com.example.tasky.authentication.domain.UserCache
+import com.example.tasky.core.data.EmailValidatorImpl
+import com.example.tasky.core.domain.EmailValidator
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
