+++ app/src/main/java/com/example/tasky/authentication/domain/SignUpFormValidator.kt
@@ -1,5 +1,6 @@
 package com.example.tasky.authentication.domain
 
+import com.example.tasky.core.domain.EmailValidator
 import javax.inject.Inject
 
 data class SignUpFormValidator @Inject constructor(
