+++ app/src/main/java/com/example/tasky/core/data/EmailValidatorImpl.kt
@@ -1,7 +1,7 @@
-package com.example.tasky.authentication.data
+package com.example.tasky.core.data
 
 import android.util.Patterns
-import com.example.tasky.authentication.domain.EmailValidator
+import com.example.tasky.core.domain.EmailValidator
 import javax.inject.Inject
 
 class EmailValidatorImpl @Inject constructor() : EmailValidator {
