+++ app/src/main/java/com/example/tasky/core/domain/EmailValidator.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.authentication.domain
+package com.example.tasky.core.domain
 
 interface EmailValidator {
     operator fun invoke(email: String): Boolean
